@@ -1,16 +1,13 @@
 name: example
 description: assets_gen examble project.
 version: 0.0.1
+publish_to: none
 
 environment:
-  sdk: ">=2.7.0 <3.0.0"
-
-dependencies:
-  sub1:
-    path: sub1/
+  sdk: ">=2.12.0 <3.0.0"
 
 dev_dependencies:
-  build_runner: ">=1.10.0 <2.0.0"
+  build_runner: ">=2.0.0 <3.0.0"
   assets_gen:
     path: ../
 
@@ -22,7 +19,6 @@ flutter:
     - assets/exclude/
     - assets/images/
     - assets/vip/
-    - assets/1.txt
     - assets2/
 
 assets_gen:
