@@ -1,19 +1,19 @@
 name: assets_gen
 description: >
   The assets_gen package helps you to generate a .dart file that contains all assets according to `pubspec.yaml`.
-version: 1.1.2
+version: 1.2.0
 repository: https://github.com/chenenyu/assets_gen
 publish_to: https://pub.dev
 
 environment:
-  sdk: ">=2.7.0 <3.0.0"
+  sdk: ">=2.12.0 <3.0.0"
 
 dependencies:
-  ansicolor: ">=1.1.0 <3.0.0"
-  args: ">=1.0.0 <3.0.0"
-  build: ">=1.0.0 <3.0.0"
-  glob: ">=1.2.0 <3.0.0"
-  logging: ">=0.11.4 <2.0.0"
-  path: ">=1.6.0 <2.0.0"
-  watcher: ">=0.9.7 <2.0.0"
-  yaml: ">=2.2.0 <4.0.0"
+  ansicolor: ">=2.0.0 <3.0.0"
+  args: ">=2.0.0 <3.0.0"
+  build: ">=2.0.0 <3.0.0"
+  glob: ">=2.0.0 <3.0.0"
+  logging: ">=1.0.0 <2.0.0"
+  path: ">=1.8.0 <2.0.0"
+  watcher: ">=1.0.0 <2.0.0"
+  yaml: ">=3.0.0 <4.0.0"
